This fixes a build failure on GCC 9.3.1, and a warning.

+++ src/core/codestream/ojph_precinct.cpp
@@ -489,7 +489,8 @@ namespace ojph {
               Lblock += bit;
             }
 
-            int bits = Lblock + 31 - count_leading_zeros(num_phld_passes + 1);
+            int bits = Lblock + 31 -
+              (int)count_leading_zeros(num_phld_passes + 1);
             if (bb_read_bits(&bb, bits, bit) == false)
             { data_left = 0; throw "error reading from file p9"; }
             if (bit < 2)
